requirements.txt
schedule==1.2.0
psycopg2-binary==2.9.9
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0
PyGithub==2.1.1
python-dotenv==1.0.0
requests==2.31.0
